Refactored packages.
LocationActivities now belong in the location package and mini-game activities in the mini-game package.
