+++ app/src/main/AndroidManifest.xml
@@ -30,22 +30,22 @@
                 android:value="nl.tudelft.pixelperfect.pixelperfect.MainActivity" />
         </activity>
         <activity
-            android:name=".LocationArmoryActivity"
+            android:name="nl.tudelft.pixelperfect.location.LocationArmoryActivity"
             android:screenOrientation="landscape"
             android:configChanges="keyboardHidden|orientation">
         </activity>
         <activity
-            android:name=".LocationEngineroomActivity"
+            android:name="nl.tudelft.pixelperfect.location.LocationEngineroomActivity"
             android:screenOrientation="landscape"
             android:configChanges="keyboardHidden|orientation">
         </activity>
         <activity
-            android:name=".LocationLabActivity"
+            android:name="nl.tudelft.pixelperfect.location.LocationLabActivity"
             android:screenOrientation="landscape"
             android:configChanges="keyboardHidden|orientation">
         </activity>
         <activity
-            android:name=".LocationDeckActivity"
+            android:name="nl.tudelft.pixelperfect.location.LocationDeckActivity"
             android:screenOrientation="landscape"
             android:configChanges="keyboardHidden|orientation">
         </activity>
