+++ app/src/main/java/nl/tudelft/pixelperfect/location/LocationEngineroomActivity.java
@@ -1,4 +1,4 @@
-package nl.tudelft.pixelperfect.pixelperfect;
+package nl.tudelft.pixelperfect.location;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
@@ -8,14 +8,16 @@ import nl.tudelft.pixelperfect.client.EventCompletedMessage;
 import nl.tudelft.pixelperfect.client.GameClient;
 import nl.tudelft.pixelperfect.event.Event;
 import nl.tudelft.pixelperfect.event.Events;
+import nl.tudelft.pixelperfect.pixelperfect.R;
+import nl.tudelft.pixelperfect.pixelperfect.Spaceship;
 
 /**
- * The location of the Janitor where he can make some coffee.
+ * This class represents the location of the Engine Room occupied by the Engineer.
  *
  * @author Floris Doolaard
  */
-@SuppressWarnings("CanBeFinal")
-public class LocationDeckActivity extends AppCompatActivity {
+@SuppressWarnings({"CanBeFinal", "unused", "UnusedParameters"})
+public class LocationEngineroomActivity extends AppCompatActivity {
     private GameClient game = GameClient.getInstance();
     private static Spaceship ship = Spaceship.getInstance();
 
@@ -27,7 +29,7 @@ public class LocationDeckActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_location_deck);
+        setContentView(R.layout.activity_location_engineroom);
     }
 
     /**
@@ -47,6 +49,19 @@ public class LocationDeckActivity extends AppCompatActivity {
     public void completeFireEvent(View view){
         if(ship.getEventLog().contains(Events.FIRE)){
             game.sendMessage(new EventCompletedMessage("Fire Event", ship.getEventLog().pop(Events.FIRE).getId()));
+        } else {
+            game.sendMessage(new EventCompletedMessage("WRONG ANSWER", -1));
+        }
+    }
+
+    /**
+     * Whenever the button to complete the Plasma Leak Event is pressed this will happen.
+     *
+     * @param view , the view of the page.
+     */
+    public void completePlasmaLeakEvent(View view){
+        if(ship.getEventLog().contains(Events.PLASMA)){
+            game.sendMessage(new EventCompletedMessage("Plasma Leak Event", ship.getEventLog().pop(Events.PLASMA).getId()));
         } else {
             game.sendMessage(new EventCompletedMessage("WRONG ANSWER", -1));
         }
