+++ app/src/main/java/nl/tudelft/pixelperfect/mini_game/AsteroidImpactActivity.java
@@ -1,0 +1,19 @@
+package nl.tudelft.pixelperfect.mini_game;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+/**
+ * The mini-game Asteroid Impact in which the use must click a wrench-button a certain amount of
+ * times to repair the ship.
+ *
+ * @author Floris Doolaard
+ */
+public class AsteroidImpactActivity extends AppCompatActivity  {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+}
